+++ types/src/libs/animation/variant/Fade/gen.d.ts
@@ -1,0 +1,4 @@
+import type { FadeProps } from '.';
+import type { GeneratedAnimationValue } from '../type';
+export declare const genAnimation: (props?: FadeProps['option'], isHover?: boolean) => GeneratedAnimationValue;
+//# sourceMappingURL=gen.d.ts.map